feat: update Stay model Localisation parameters to Long[]

+++ services/stayservice/src/main/java/com/woofwoof/stayservice/models/Stay.java
@@ -20,8 +20,7 @@ public class Stay {
 
     private String description;
 
-    private int[] localisation;
-    //private Long[] localisation;
+    private Long[] localisation;
 
     private Date startDate;
